confirmation dialog

+++ src/group/group-list/group-list.component.ts
@@ -1,12 +1,15 @@
-import { Component } from '@angular/core';
+import { Component, TemplateRef, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Select, Store } from '@ngxs/store';
 import { Observable, take, tap } from 'rxjs';
 import { GroupsService } from 'src/api/groups.service';
 import { GroupRole } from 'src/enums/group-role.enum';
 import { Group } from 'src/models';
 import { SnackbarService } from 'src/services/snackbar.service';
+import { ConfirmationDialogComponent } from 'src/shared-ui/confirmation-dialog/confirmation-dialog.component';
 import { GroupState } from 'src/store/group.state';
 import { RemoveGroup } from 'src/store/group.state.actions';
+import { DEFAULT_DIALOG_CONFIG } from '../../../constants';
 
 @Component({
   selector: 'app-group-list',
@@ -21,7 +24,8 @@ export class GroupListComponent {
   constructor(
     private groupsService: GroupsService,
     private store: Store,
-    private snackbarService: SnackbarService
+    private snackbarService: SnackbarService,
+    private matDialog: MatDialog
   ) {}
 
   public displayedColumns = [
@@ -37,16 +41,28 @@ export class GroupListComponent {
     const groups = this.store.selectSnapshot(GroupState.groups);
     if (groups.length > 1) {
       const group = groups[index];
-      this.groupsService
-        .deleteGroup(group.id.toString())
-        .pipe(
-          take(1),
-          tap(() => {
-            this.snackbarService.success('Group successfully deleted');
-            this.store.dispatch(new RemoveGroup(group.id.toString()));
-          })
-        )
-        .subscribe();
+      const dialogRef = this.matDialog.open(
+        ConfirmationDialogComponent,
+        DEFAULT_DIALOG_CONFIG
+      );
+
+      dialogRef.componentInstance.headerText = 'Delete Group';
+      dialogRef.componentInstance.dialogContent = `Are you sure you would like to the group '${group.name}'? All receipts will be deleted as a result.`;
+
+      dialogRef.afterClosed().subscribe((r) => {
+        if (r) {
+          this.groupsService
+            .deleteGroup(group.id.toString())
+            .pipe(
+              take(1),
+              tap(() => {
+                this.snackbarService.success('Group successfully deleted');
+                this.store.dispatch(new RemoveGroup(group.id.toString()));
+              })
+            )
+            .subscribe();
+        }
+      });
     }
   }
 }
